requirements.txt
Django==5.1.5
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-decouple==3.8
django-debug-toolbar==5.0.1
pillow==10.4.0
flake8==7.1.1
pytest==8.3.4
pytest-django==4.9.0 
pytest-cov==6.0.0
dj-database-url==2.3.0
whitenoise==6.8.2
cloudinary==1.42.1
django-cloudinary-storage==0.3.0
django-ckeditor==6.7.2
openai==1.63.2
requests==2.32.3